fura to master, working fine

+++ ibay/Controllers/UserController.cs
@@ -1,8 +1,11 @@
-﻿using ibay.Model;
+﻿using System.Security.Claims;
+using ibay.Model;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using ibay.Filters;
 using ibay.Services;
+using Microsoft.AspNetCore.Authorization;
+using Npgsql.EntityFrameworkCore.PostgreSQL.Infrastructure.Internal;
 
 
 namespace ibay.Controllers
@@ -11,7 +14,7 @@ namespace ibay.Controllers
     [ApiController]
     [ExecutedReqFilter]
 
-    public class UserController: ControllerBase
+    public class UserController : ControllerBase
     {
         [HttpGet]
         [Route("getby")]
@@ -19,7 +22,10 @@ namespace ibay.Controllers
         {
             var id = itemId.Id;
             var User = ibay.GetUserById(id);
-            if (User == null) { return NotFound($"User id : {id} not found"); }
+            if (User == null)
+            {
+                return NotFound($"User id : {id} not found");
+            }
 
             return Ok(User);
         }
@@ -42,23 +48,29 @@ namespace ibay.Controllers
 
         [HttpPut]
         [Route("update")]
-        public IActionResult Put(IIbay ibay, [FromBody] User User)
+        [Authorize]
+        public IActionResult Put(IIbay ibay, [FromBody] User user)
         {
-            ibay.UpdateUser(User.Id, User);
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId != user.Id.ToString()) return Unauthorized("You are not allowed to modify other users");
+            ibay.UpdateUser(user.Id, user);
             return Ok();
         }
 
         [HttpDelete]
         [Route("delete")]
+        [Authorize]
         public IActionResult Delete(IIbay ibay, [FromBody] ItemId itemId)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId != itemId.Id.ToString()) return Unauthorized("You are not allowed to delete other users");
             var id = itemId.Id;
             ibay.DeleteUser(id);
             return Ok();
         }
 
 
-        
+
     }
 
     [Route("/api/product/")]
@@ -89,17 +101,24 @@ namespace ibay.Controllers
         {
             var id = itemId.Id;
             var Product = ibay.GetProductById(id);
-            if (Product == null) { return NotFound($"Product Id : {id} not found"); }
+            if (Product == null)
+            {
+                return NotFound($"Product Id : {id} not found");
+            }
 
             return Ok(Product);
         }
 
         [HttpPost]
         [Route("new")]
+        [Authorize]
         public IActionResult Post(IIbay ibay, [FromBody] Product product)
         {
+            var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (userRole != "seller" && userRole != "admin") return Unauthorized("You are not allowed to Modify products");
             try
             {
+                product.SellerId = new Guid(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
                 var id = ibay.CreateProduct(product);
 
                 return Created($"/api/stud/{id}", id);
@@ -112,23 +131,133 @@ namespace ibay.Controllers
 
         [HttpPut]
         [Route("update")]
+        [Authorize]
         public IActionResult Put(IIbay ibay, [FromBody] Product product)
         {
+            var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (userRole != "seller" && userRole != "admin") return Unauthorized("You are not allowed to Modify products");
+            
+            var checkProduct = ibay.GetProductById(product.Id);
+            if (checkProduct == null)
+            {
+                return NotFound($"Product Id : {product.Id} not found");
+            }
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId != checkProduct.SellerId.ToString())
+                return Unauthorized("You are only allowed to Modify your products");
+            
             ibay.UpdateProduct(product.Id, product);
             return Ok();
         }
 
         [HttpDelete]
         [Route("delete")]
+        [Authorize]
         public IActionResult Delete(IIbay ibay, [FromBody] ItemId itemId)
         {
             var id = itemId.Id;
+            var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (userRole != "seller" && userRole != "admin") return Unauthorized("You are not allowed to Modify products");
+            
+            var checkProduct = ibay.GetProductById(id);
+            if (checkProduct == null)
+            {
+                return NotFound($"Product Id : {id} not found");
+            }
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId != checkProduct.SellerId.ToString())
+                return Unauthorized("You are only allowed to Modify your products");
+            
             ibay.DeleteProduct(id);
             return Ok();
         }
 
 
+
+    }
+
+    [Route("/api/cart/")]
+    [ApiController]
+    [ExecutedReqFilter]
+
+    public class CartController : ControllerBase
+    {
+        [HttpPost]
+        [Authorize]
+        [Route("add")]
+
+        public IActionResult Add(IIbay ibay, [FromBody] Cart cart)
+        {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            cart.UserId = new Guid(userId);
+            ibay.AddToCart(cart);
+            return Ok();
+        }
         
+        [HttpDelete]
+        [Authorize]
+        [Route("remove")]
+
+        public IActionResult remove(IIbay ibay, [FromBody] ItemId cart)
+        {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            ibay.RemoveFromCart(userId,cart.Id);
+            return Ok();
+        }
+        
+        [HttpDelete]
+        [Authorize]
+        [Route("empty")]
+        public IActionResult EmptyCart(IIbay ibay)
+        {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            ibay.EmptyCart(userId);
+            return Ok();
+        }
+        
+        [HttpPost]
+        [Authorize]
+        [Route("pay")]
+        public IActionResult PayCart(IIbay ibay)
+        {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            ibay.EmptyCart(userId);
+            return Ok();
+        }
+        
+        [HttpGet]
+        [Authorize]
+        [Route("list")]
+        public IActionResult GetCartItems(IIbay ibay)
+        {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var cartItems = ibay.GetCartItems(userId);
+            return Ok(cartItems);
+        }
+    }
+
+    [Route("/api/auth/")]
+    [ApiController]
+    [ExecutedReqFilter]
+    public class AuthController : ControllerBase
+    {
+        [HttpPost]
+        [Route("login")]
+        public IActionResult Login(IJwtAuthService jwtAuthService, [FromBody] LoginModel login)
+        {
+            var user = jwtAuthService.Authenticate(login.Username, login.Password);
+            if (user == null) return Unauthorized("User not found, wrong username/password");
+
+            var claims = new List<Claim>
+            {
+                new Claim(ClaimTypes.NameIdentifier, user.Id.ToString()),
+                new Claim(ClaimTypes.Role, user.Role)
+            };
+            var token = jwtAuthService.GenerateToken(
+                "MON6SUPER6SECRET6EST6UN6GRAND6MOT6DE6PASSE6DUNE6TRENTAINE6DE6CARACTERES", claims);
+
+            return Ok(token);
+        }
     }
 }
